% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{count_multiple}
\alias{count_multiple}
\title{helper function}
\usage{
count_multiple(data, question_analysis, group = NULL, sort = TRUE)
}
\arguments{
\item{data}{a dataset}

\item{question_analysis}{the main variable in the analysis}

\item{group}{the variable to use as the grouping variable}

\item{sort}{the variable to use as the sorting variable}
}
\description{
function to count the values of single-choice questions
}
