% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{extract_labels}
\alias{extract_labels}
\title{helper function}
\usage{
extract_labels(data, group = NULL)
}
\arguments{
\item{data}{a dataset}

\item{group}{the variable to use as the grouping variable}
}
\description{
function to count the variables of interest
}
