% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{load_text}
\alias{load_text}
\title{helper function}
\usage{
load_text(question_analysis, group = NULL)
}
\arguments{
\item{question_analysis}{a dataset}

\item{group}{the variable to use as the grouping variable}
}
\description{
function to count the variables of interest
}
