% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{plot_q3_2}
\alias{plot_q3_2}
\title{helper function}
\usage{
plot_q3_2(country = NULL)
}
\arguments{
\item{country}{filter analysis by country}
}
\description{
function to prepare the variables of interest
}
