% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_03_demographics_fct_helpers.R
\name{plot_q3_5}
\alias{plot_q3_5}
\title{helper function}
\usage{
plot_q3_5(country = NULL)
}
\arguments{
\item{country}{plot age distribution}
}
\description{
what are respondents' marital status?
}
