% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_03_demographics_fct_helpers.R
\name{plot_q3_6}
\alias{plot_q3_6}
\title{helper function}
\usage{
plot_q3_6(country = NULL)
}
\arguments{
\item{country}{plot age distribution}
}
\description{
what languages do respondents speak?
}
