% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_03_demographics_fct_helpers.R
\name{plot_q3_9}
\alias{plot_q3_9}
\title{helper function}
\usage{
plot_q3_9(country = NULL)
}
\arguments{
\item{country}{area of study}
}
\description{
function to prepare the variables of interest
}
