% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_04_experience_fct_helpers.R
\name{plot_q4_4}
\alias{plot_q4_4}
\title{helper function}
\usage{
plot_q4_4(country = NULL)
}
\arguments{
\item{country}{what were they doing before leaving and currently}
}
\description{
function to prepare the variables of interest
}
