% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_03_demographics_fct_helpers.R
\name{plot_q4_7}
\alias{plot_q4_7}
\title{helper function}
\usage{
plot_q4_7(country = NULL)
}
\arguments{
\item{country}{why settle}
}
\description{
function to prepare the variables of interest
}
