% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_04_experience_fct_helpers.R
\name{plot_q5_12}
\alias{plot_q5_12}
\alias{plot_q5_11}
\title{helper function}
\usage{
plot_q5_11(country = NULL)

plot_q5_12(country = NULL)
}
\arguments{
\item{country}{entrepreneurship background}
}
\description{
function to prepare the variables of interest

function to prepare the variables of interest
}
