% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_selection_fct_helpers.R
\name{plot_q5_2}
\alias{plot_q5_2}
\title{helper function}
\usage{
plot_q5_2(country = NULL)
}
\arguments{
\item{country}{what were they doing before leaving and currently}
}
\description{
function to prepare the variables of interest
}
