% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_04_experience_fct_helpers.R
\name{plot_q5_3}
\alias{plot_q5_3}
\title{helper function}
\usage{
plot_q5_3(country = NULL)
}
\arguments{
\item{country}{plot time of residency}
}
\description{
function to prepare the variables of interest
}
