% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_04_experience_fct_helpers.R
\name{plot_q5_8}
\alias{plot_q5_8}
\title{helper function}
\usage{
plot_q5_8(country = NULL)
}
\arguments{
\item{country}{what educational leve are their pursuing}
}
\description{
function to prepare the variables of interest
}
