% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_04_experience_fct_helpers.R
\name{plot_q5_9}
\alias{plot_q5_9}
\title{helper function}
\usage{
plot_q5_9(country = NULL)
}
\arguments{
\item{country}{what area of study are they in?}
}
\description{
function to prepare the variables of interest
}
