% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_05_attachment_fct_helpers.R
\name{plot_q6_5}
\alias{plot_q6_5}
\title{helper function}
\usage{
plot_q6_5(country = NULL)
}
\arguments{
\item{country}{filter by country}
}
\description{
how often respondents talk to people in Colombia?
}
