% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_05_attachment_fct_helpers.R
\name{plot_q6_6}
\alias{plot_q6_6}
\title{helper function}
\usage{
plot_q6_6(country = NULL)
}
\arguments{
\item{country}{filter by country}
}
\description{
do respondents have pension funds in Colombia?
}
