% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_06_return_fct_helpers.R
\name{plot_q7_2}
\alias{plot_q7_2}
\title{helper function}
\usage{
plot_q7_2(country = NULL)
}
\arguments{
\item{country}{plot family members or friends who live in colombia}
}
\description{
function to prepare the variables of interest
}
