% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_UI}
\alias{plot_UI}
\title{dinamically produce UI}
\usage{
plot_UI(number, ns)
}
\arguments{
\item{number}{a number indicating the element of the UI}
}
\description{
function to count the variables of interest
}
