% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{prepare_data}
\alias{prepare_data}
\title{helper function}
\usage{
prepare_data(
  question_analysis = NULL,
  filter = c(q3_2 = NULL),
  extra_filter = c(`NULL` = NULL)
)
}
\arguments{
\item{question_analysis}{the main variable in the analysis}

\item{filter}{the variable to filter the data by}

\item{extra_filter}{an additional filter passed onto the dataset}
}
\description{
function to prepare the variables of interest
}
