% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_selection_fct_helpers.R
\name{prepare_plot_q136}
\alias{prepare_plot_q136}
\alias{plot_q136}
\title{helper function}
\usage{
plot_q136(country = NULL)
}
\arguments{
\item{country}{filter analysis by country}
}
\description{
where do they live
}
