% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_selection_fct_helpers.R
\name{prepare_plot_q2_3}
\alias{prepare_plot_q2_3}
\alias{plot_q2_3}
\title{helper function}
\usage{
plot_q2_3(country = NULL)
}
\arguments{
\item{country}{filter analysis by country}
}
\description{
function to prepare the variables of interest
}
