% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_06_return_fct_helpers.R
\name{prepare_plot_q7_14}
\alias{prepare_plot_q7_14}
\alias{plot_q7_14}
\title{helper function}
\usage{
plot_q7_14(country = NULL)
}
\arguments{
\item{country}{filter analysis by country}
}
\description{
relevant policy for family reasons
}
