% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_06_return_fct_helpers.R
\name{prepare_plot_q7_8}
\alias{prepare_plot_q7_8}
\alias{plot_q7_8}
\title{helper function}
\usage{
plot_q7_8(country = NULL)
}
\arguments{
\item{country}{filter analysis by country}
}
\description{
engagement with colombia for people who returned due to COVID
}
