% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12058 rows and 103 columns.
}
\usage{
survey_data
}
\description{
data
}
\keyword{datasets}
