% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analysis_text}
\alias{analysis_text}
\title{analysis text}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 7 columns.
}
\usage{
analysis_text
}
\description{
analysis text
}
\keyword{datasets}
